sync_script/requirements.txt
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
dj-config-url==0.1.1
Django==4.0.5
django-debug-toolbar==3.4.0
django-gsheets==0.0.10
django-stubs==1.12.0
django-stubs-ext==0.5.0
fab==3.1.0
Fabric3==1.14.post1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-api-python-client-stubs==1.11.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
mypy==0.961
mypy-extensions==0.4.3
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
paramiko==2.11.0
Pillow==9.1.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
types-psycopg2==2.9.16
types-pytz==2021.3.8
types-PyYAML==6.0.8
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9

